Re-create `Model` class in Job.py for each resource type
