+++ tests/unit/scheduler_test/test_local_scheduler.py
@@ -1,5 +1,5 @@
 from aiaccel.scheduler.local_scheduler import LocalScheduler
-
+from aiaccel.scheduler.job.job import Job
 from tests.base_test import BaseTest
 
 
@@ -27,7 +27,10 @@ class TestLocalScheduler(BaseTest):
                 "10 00:00:10 2020\n"
             ]
         )
-        scheduler.jobs.append({'trial_id': 'sample1', 'thread': None})
+
+        trial_id = 1
+        job = Job(self.config, scheduler, trial_id) 
+        scheduler.jobs.append(job)
         assert scheduler.get_stats() is None
 
     def test_parse_trial_id(self, config_json):
